shell: refactor the code into smaller pieces
